Zero-downtime deployments
Issue #10

When deploying a new version of an app, ensures traffic moves from the
previous version with no service interruption. This is done by waiting
until new instances are `InService`, and then removing the previous
instances (this means for a brief period both the new version and
previous version receive traffic).

Also ensures the same number of instances in the load balancer between
deployments.